cli/requirements.txt

GitPython==3.1.43
svgwrite==1.4.1
